Count junctions

Count junctions from either TopHat2, HISAT2 or Rail-RNA output

junctionCount(
  junctionFiles,
  sampleNames = names(junctionFiles),
  output = c("Count", "Rail"),
  minOverhang = 0,
  strandSpecific = FALSE,
  illuminaStranded = FALSE,
  minCount = 1,
  maxCores = 1,
  skipLines = ifelse(output == "Count", 0, 1)
)

Arguments

junctionFiles

A character vector with the full paths to the junction files. Can alternatively be a list of GRanges-class objects with the junction count information.

sampleNames

A character vector of the same length as junctionsFiles to use as the sample names.

output

Either Count (TopHat2, HISAT2) or Rail (Rail-RNA).

minOverhang

minimum overhang.

strandSpecific

a logical specifying whether the library is strand specific or not.

illuminaStranded

a logical indicating whether the Illumina library is stranded or not.

minCount

Minimum count.

maxCores

The maximum number of cores to use. By default one.

skipLines

An integer(1) specifying how many lines to skip on the junction files. Some files have a header and some don't. So you'll either want to use 1 or 0 in most scenarios.

Value

A two element list with a DataFrame and a GRanges-class object with the counts and the annotation used.

Author

Andrew E Jaffe

Examples

## TODO (need some data)
## See https://github.com/LieberInstitute/RNAseq-pipeline/blob/ab71dedb36bcc3dad57233e645fabd5deb96d446/sh/create_count_objects-human.R#L633-L643
## for an example of how it's being using the the RNA-seq pipeline code