Merge sample metadata across several columns

This function merges the sample metadata from a RangedSummarizedExperiment object for some of the variables that we (at Jaffelab) typically create with our RNA-seq processing pipeline for samples that are sequenced in more than one lane.

merge_rse_metrics(rse)

Arguments

rse
A RangedSummarizedExperiment-class object with 'concordMapRate', 'overallMapRate', 'mitoRate', 'rRNA_rate', 'totalAssignedGene', 'numMapped', 'numReads', 'numUnmapped', 'mitoMapped', 'totalMapped' as either NumericList() or IntegerList() objects in the colData() columns.

Value

A RangedSummarizedExperiment-class with merged columns for 'concordMapRate', 'overallMapRate', 'mitoRate', 'rRNA_rate', 'totalAssignedGene', 'numMapped', 'numReads', 'numUnmapped', 'mitoMapped', 'totalMapped'.

Author

Leonardo Collado-Torres, Andrew E Jaffe

Examples

## Taken from ?SummarizedExperiment
library('SummarizedExperiment')

#> Loading required package: MatrixGenerics

#> Loading required package: matrixStats

#> 
#> Attaching package: ‘MatrixGenerics’

#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars

#> Loading required package: GenomicRanges

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following object is masked from ‘package:limma’:
#> 
#>     plotMA

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which.max, which.min

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following objects are masked from ‘package:base’:
#> 
#>     expand.grid, I, unname

#> Loading required package: IRanges

#> Loading required package: GenomeInfoDb

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> 
#> Attaching package: ‘Biobase’

#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians

#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians

nrows <- 200; ncols <- 6
set.seed(20181116)
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
rowRanges <- GRanges(rep(c("chr1", "chr2"), c(50, 150)),
                     IRanges(floor(runif(200, 1e5, 1e6)), width=100),
                     strand=sample(c("+", "-"), 200, TRUE),
                     feature_id=sprintf("ID%03d", 1:200))

## Function for generating some data
rand_gen <- function(mu = 1e6, sd = 1e3, digits = 0) {
    IRanges::NumericList(lapply(1:6, function(x) {
        round(rnorm(2, mu, sd), digits)
    }))
}
colData <- DataFrame(Treatment=rep(c("ChIP", "Input"), 3),
                     row.names=LETTERS[1:6],
                     concordMapRate = rand_gen(0.5, 0.05, 10),
                     overallMapRate = rand_gen(0.7, 0.05, 10),
                     mitoRate = rand_gen(0.2, 0.05, 10),
                     rRNA_rate = rand_gen(0.1, 0.01, 10),
                     totalAssignedGene = rand_gen(0.5, 0.05, 10),
                     numMapped = rand_gen(),
                     numReads = rand_gen(1e6 + 6e5),
                     numUnmapped = rand_gen(5e5),
                     mitoMapped = rand_gen(1e5),
                     totalMapped = rand_gen(1e6 + 1e5)
)
rse <- SummarizedExperiment(assays=SimpleList(counts=counts),
                            rowRanges=rowRanges, colData=colData)
colData(rse)

#> DataFrame with 6 rows and 11 columns
#>     Treatment    concordMapRate    overallMapRate          mitoRate
#>   <character>     <NumericList>     <NumericList>     <NumericList>
#> A        ChIP 0.541645,0.581919 0.638841,0.744259 0.196154,0.176830
#> B       Input 0.636704,0.417308 0.720357,0.808207 0.206167,0.307936
#> C        ChIP 0.480932,0.563030 0.766605,0.731026 0.263311,0.191516
#> D       Input 0.514338,0.404294 0.680493,0.706884 0.199656,0.201472
#> E        ChIP 0.575868,0.580396 0.675552,0.721660 0.283546,0.200031
#> F       Input 0.487973,0.549464 0.758194,0.680816 0.254013,0.122518
#>             rRNA_rate totalAssignedGene       numMapped        numReads
#>         <NumericList>     <NumericList>   <NumericList>   <NumericList>
#> A 0.0982325,0.1004535 0.534906,0.495279  999489,1001045 1600514,1599856
#> B   0.108117,0.105019 0.512784,0.585948   999904,999645 1600861,1600902
#> C 0.0935756,0.1128849 0.500233,0.543947 1000994, 999994 1599045,1599113
#> D 0.0983786,0.0961797 0.454453,0.394858 1000080,1001643 1600388,1601137
#> E 0.0786774,0.0902039 0.511399,0.444904   999405,998558 1598587,1600106
#> F   0.116688,0.101786 0.512436,0.581784 1000598, 999127 1598227,1598982
#>     numUnmapped    mitoMapped     totalMapped
#>   <NumericList> <NumericList>   <NumericList>
#> A 498799,500663 100075, 99746 1100621,1098107
#> B 499160,501936 100454,100220 1099005,1100717
#> C 499887,499965 100077,100616 1099423,1101197
#> D 499315,499551   98586,97819 1098706,1098371
#> E 499205,500651  98690,100203 1099373,1101728
#> F 500096,501025   99278,98614 1101246,1098881

rse_merged <- merge_rse_metrics(rse)
colData(rse_merged)

#> DataFrame with 6 rows and 11 columns
#>     Treatment concordMapRate overallMapRate  mitoRate rRNA_rate
#>   <character>      <numeric>      <numeric> <numeric> <numeric>
#> A        ChIP       0.561778       0.691539  0.186484 0.0993439
#> B       Input       0.527004       0.764283  0.257045 0.1065683
#> C        ChIP       0.521982       0.748815  0.227432 0.1032254
#> D       Input       0.459303       0.693692  0.200565 0.0972783
#> E        ChIP       0.578133       0.698617  0.241806 0.0844382
#> F       Input       0.518726       0.719496  0.188314 0.1092428
#>   totalAssignedGene numMapped  numReads numUnmapped mitoMapped totalMapped
#>           <numeric> <numeric> <numeric>   <numeric>  <numeric>   <numeric>
#> A          0.515077   2000534   3200370      999462     199821     2198728
#> B          0.549361   1999549   3201763     1001096     200674     2199722
#> C          0.522079   2000988   3198158      999852     200693     2200620
#> D          0.424632   2001723   3201525      998866     196405     2197077
#> E          0.478166   1997963   3198693      999856     198893     2201101
#> F          0.547085   1999725   3197209     1001121     197892     2200127