Merge sample metadata across several columns

This function merges the sample metadata from a RangedSummarizedExperiment object for some of the variables that we (at Jaffelab) typically create with our RNA-seq processing pipeline for samples that are sequenced in more than one lane.

merge_rse_metrics(rse)

Arguments

rse
A RangedSummarizedExperiment-class object with 'concordMapRate', 'overallMapRate', 'mitoRate', 'rRNA_rate', 'totalAssignedGene', 'numMapped', 'numReads', 'numUnmapped', 'mitoMapped', 'totalMapped' as either NumericList() or IntegerList() objects in the colData() columns.

Value

A RangedSummarizedExperiment-class with merged columns for 'concordMapRate', 'overallMapRate', 'mitoRate', 'rRNA_rate', 'totalAssignedGene', 'numMapped', 'numReads', 'numUnmapped', 'mitoMapped', 'totalMapped'.

Examples

## Taken from ?SummarizedExperiment
library('SummarizedExperiment')

#> Loading required package: GenomicRanges

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following object is masked from ‘package:limma’:
#> 
#>     plotMA

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which, which.max, which.min

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> Loading required package: GenomeInfoDb

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: DelayedArray

#> Loading required package: matrixStats

#> 
#> Attaching package: ‘matrixStats’

#> The following objects are masked from ‘package:Biobase’:
#> 
#>     anyMissing, rowMedians

#> Loading required package: BiocParallel

#> 
#> Attaching package: ‘DelayedArray’

#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colMaxs, colMins, colRanges, rowMaxs, rowMins, rowRanges

#> The following objects are masked from ‘package:base’:
#> 
#>     aperm, apply, rowsum

nrows <- 200; ncols <- 6
set.seed(20181116)
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
rowRanges <- GRanges(rep(c("chr1", "chr2"), c(50, 150)),
                     IRanges(floor(runif(200, 1e5, 1e6)), width=100),
                     strand=sample(c("+", "-"), 200, TRUE),
                     feature_id=sprintf("ID%03d", 1:200))

## Function for generating some data
rand_gen <- function(mu = 1e6, sd = 1e3, digits = 0) {
    IRanges::NumericList(lapply(1:6, function(x) {
        round(rnorm(2, mu, sd), digits)
    }))
}
colData <- DataFrame(Treatment=rep(c("ChIP", "Input"), 3),
                     row.names=LETTERS[1:6],
                     concordMapRate = rand_gen(0.5, 0.05, 10),
                     overallMapRate = rand_gen(0.7, 0.05, 10),
                     mitoRate = rand_gen(0.2, 0.05, 10),
                     rRNA_rate = rand_gen(0.1, 0.01, 10),
                     totalAssignedGene = rand_gen(0.5, 0.05, 10),
                     numMapped = rand_gen(),
                     numReads = rand_gen(1e6 + 6e5),
                     numUnmapped = rand_gen(5e5),
                     mitoMapped = rand_gen(1e5),
                     totalMapped = rand_gen(1e6 + 1e5)
)
rse <- SummarizedExperiment(assays=SimpleList(counts=counts),
                            rowRanges=rowRanges, colData=colData)
colData(rse)

#> DataFrame with 6 rows and 11 columns
#>     Treatment            concordMapRate            overallMapRate
#>   <character>             <NumericList>             <NumericList>
#> A        ChIP 0.5416450277,0.5819190164 0.6388409556,0.7442588758
#> B       Input 0.6367038563,0.4173075052 0.7203572166,0.8082071374
#> C        ChIP 0.4809315501,0.5630297479 0.7666049766,0.7310258555
#> D       Input 0.5143381213,0.4042940316 0.6804930318,0.7068839356
#> E        ChIP 0.5758675543,0.5803962241 0.6755520331,0.7216600205
#> F       Input  0.4879732694,0.549463787 0.7581941353,0.6808157727
#>                    mitoRate                 rRNA_rate         totalAssignedGene
#>               <NumericList>             <NumericList>             <NumericList>
#> A 0.1961535681,0.1768301236 0.0982325258,0.1004534757 0.5349062104,0.4952787598
#> B 0.2061674148,0.3079357627 0.1081174562,0.1050187044 0.5127843394,0.5859475281
#> C 0.2633111231,0.1915161119 0.0935755694,0.1128848613 0.5002328801,0.5439469979
#> D 0.1996563385,0.2014715116 0.0983786493,0.0961797209 0.4544526261,0.3948582872
#> E   0.28354641,0.2000309501 0.0786773534,0.0902038669 0.5113993705,0.4449038444
#> F 0.2540125234,0.1225183117 0.1166883294,0.1017862296 0.5124364307,0.5817841252
#>         numMapped        numReads   numUnmapped    mitoMapped     totalMapped
#>     <NumericList>   <NumericList> <NumericList> <NumericList>   <NumericList>
#> A  999489,1001045 1600514,1599856 498799,500663  100075,99746 1100621,1098107
#> B   999904,999645 1600861,1600902 499160,501936 100454,100220 1099005,1100717
#> C  1000994,999994 1599045,1599113 499887,499965 100077,100616 1099423,1101197
#> D 1000080,1001643 1600388,1601137 499315,499551   98586,97819 1098706,1098371
#> E   999405,998558 1598587,1600106 499205,500651  98690,100203 1099373,1101728
#> F  1000598,999127 1598227,1598982 500096,501025   99278,98614 1101246,1098881

rse_merged <- merge_rse_metrics(rse)
colData(rse_merged)

#> DataFrame with 6 rows and 11 columns
#>     Treatment    concordMapRate    overallMapRate          mitoRate
#>   <character>         <numeric>         <numeric>         <numeric>
#> A        ChIP 0.561777881859246 0.691539078673111 0.186484331036545
#> B       Input 0.527004276016046 0.764282739478665 0.257044997763206
#> C        ChIP 0.521981521795755 0.748815037804111 0.227431557390494
#> D       Input  0.45930320398716 0.693691570789269 0.200564633718371
#> E        ChIP 0.578132964490662 0.698616974714173 0.241806382478557
#> F       Input 0.518725788491834  0.71949581780281 0.188313781196354
#>            rRNA_rate totalAssignedGene numMapped  numReads numUnmapped
#>            <numeric>         <numeric> <numeric> <numeric>   <numeric>
#> A 0.0993438644688982 0.515077074136444   2000534   3200370      999462
#> B  0.106568280989435 0.549361195365026   1999549   3201763     1001096
#> C  0.103225390410545 0.522079015866578   2000988   3198158      999852
#> D 0.0972783266083183 0.424632190206116   2001723   3201525      998866
#> E  0.084438166922339 0.478165702233188   1997963   3198693      999856
#> F  0.109242760500839 0.547084771828256   1999725   3197209     1001121
#>   mitoMapped totalMapped
#>    <numeric>   <numeric>
#> A     199821     2198728
#> B     200674     2199722
#> C     200693     2200620
#> D     196405     2197077
#> E     198893     2201101
#> F     197892     2200127